Make ContentDiaglog use standard buttons
