+++ tools/QuietBackgroundProcesses/DevHome.QuietBackgroundProcesses.UI/Views/AnalyticSummaryPopup.xaml
@@ -8,27 +8,17 @@
     xmlns:local="using:DevHome.QuietBackgroundProcesses.UI.Views"
     xmlns:i="using:Microsoft.Xaml.Interactivity"
     xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    x:Uid="AnalyticSummaryContentDialog"
+    PrimaryButtonClick="SaveReportButtonClicked"
     Style="{StaticResource DefaultContentDialogStyle}">
 
-    <!-- ContentDialog Width and Height are not properly hooked up and must be set this way -->
-    <ContentDialog.Resources>
-        <x:Double x:Key="ContentDialogMinWidth">600</x:Double>
-        <x:Double x:Key="ContentDialogMaxWidth">600</x:Double>
-        <x:Double x:Key="ContentDialogMaxHeight">590</x:Double>
-        <Thickness x:Key="ContentDialogTitleMargin">0,0,0,0</Thickness>
-        <Thickness x:Key="ContentDialogPadding">0,0,0,0</Thickness>
-        <Thickness x:Key="NavigationViewContentMargin">0,0,0,0</Thickness>
-    </ContentDialog.Resources>
-
-    <!-- Title and Close button -->
-    <Grid Height="600">
+    <Grid Height="480" Width="500">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
 
-        <StackPanel Padding="15">
+        <StackPanel Grid.Row="0">
             <TextBlock
                 x:Uid="AnalyticSummary_Title"
                 Style="{StaticResource SubtitleTextBlockStyle}"
@@ -82,46 +72,9 @@
             </Grid>
         </StackPanel>
 
-        <ScrollViewer VerticalScrollBarVisibility="Auto" Grid.Row="1" Padding="10 0 10 0">
-            <StackPanel Padding="15 0 15 15">
-                <local:ProcessPerformanceTableControl x:Name="processListControl" ItemsSource="{x:Bind ViewModel.ProcessDatasAd}" />
-            </StackPanel>
+        <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto" Margin="0 20 0 0">
+            <local:ProcessPerformanceTableControl x:Name="processListControl" ItemsSource="{x:Bind ViewModel.ProcessDatasAd}" />
         </ScrollViewer>
-
-        <Grid Grid.Row="2" Background="{StaticResource SmokeFillColorDefaultBrush}">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition />
-                <ColumnDefinition />
-            </Grid.ColumnDefinitions>
-
-            <Button
-                x:Name="SaveButton"
-                x:Uid="SaveButton"
-                Width="Auto"
-                Command="{x:Bind ViewModel.SaveReportButtonClickedCommand}"
-                Style="{ThemeResource AccentButtonStyle}"
-                VerticalAlignment="Bottom"
-                HorizontalAlignment="Stretch"
-                HorizontalContentAlignment="Center"
-                Margin="20,20,5,35"
-                MinHeight="32" MinWidth="118">
-                <TextBlock x:Uid="AnalyticSummary_SaveReportButton" FontSize="14" />
-            </Button>
-
-            <Button
-                x:Name="CloseButton"
-                x:Uid="CloseButton"
-                Grid.Column="1"
-                Width="Auto"
-                Click="CloseButton_Click"
-                VerticalAlignment="Bottom"
-                HorizontalAlignment="Stretch"
-                HorizontalContentAlignment="Center"
-                Margin="5,20,20,35"
-                MinHeight="32" MinWidth="118">
-                <TextBlock x:Uid="AnalyticSummary_CloseButton" FontSize="14" />
-            </Button>
-        </Grid>
     </Grid>
 
 </ContentDialog>
